+++ UI/src/util/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_PAGE_SIZE = 20;
